# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/custom-fields": {
      "post": {
        "operationId": "CustomFieldsController_create",
        "summary": "Create a new custom field",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomFieldDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The custom field has been successfully created."
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "tags": [
          "custom-fields"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "CustomFieldsController_findAll",
        "summary": "Get all custom fields",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all custom fields."
          },
          "404": {
            "description": "Custom fields not found."
          }
        },
        "tags": [
          "custom-fields"
        ]
      }
    },
    "/custom-fields/{id}": {
      "get": {
        "operationId": "CustomFieldsController_findOne",
        "summary": "Get a custom field by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return a custom field."
          },
          "404": {
            "description": "Custom field not found."
          }
        },
        "tags": [
          "custom-fields"
        ]
      },
      "put": {
        "operationId": "CustomFieldsController_update",
        "summary": "Update a custom field by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomFieldDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The custom field has been successfully updated."
          },
          "404": {
            "description": "Custom field not found."
          }
        },
        "tags": [
          "custom-fields"
        ]
      },
      "delete": {
        "operationId": "CustomFieldsController_remove",
        "summary": "Delete a custom field by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The custom field has been successfully deleted."
          },
          "404": {
            "description": "Custom field not found."
          }
        },
        "tags": [
          "custom-fields"
        ]
      }
    },
    "/custom-fields/me/{schema}": {
      "get": {
        "operationId": "CustomFieldsController_findAllOwnerCustomsFields",
        "summary": "Get all custom fields for the current owner and schema",
        "parameters": [
          {
            "name": "schema",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all custom fields for the current owner and schema."
          },
          "404": {
            "description": "Custom fields not found."
          }
        },
        "tags": [
          "custom-fields"
        ]
      },
      "put": {
        "operationId": "CustomFieldsController_updatePositions",
        "summary": "Update positions of custom fields for the current owner and schema",
        "parameters": [
          {
            "name": "schema",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The positions of the custom fields have been successfully updated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "custom-fields"
        ]
      }
    },
    "/custom-fields/{id}/me/{schema}": {
      "get": {
        "operationId": "CustomFieldsController_findOneOwnerCustomsFields",
        "summary": "Get a specific custom field for the current owner and schema",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "schema",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the custom field for the current owner and schema."
          },
          "404": {
            "description": "Custom field not found."
          }
        },
        "tags": [
          "custom-fields"
        ]
      }
    },
...
  },
  "info": {
    "title": "Api ZenFlow",
    "description": "Api for a CRM/Forum application with modules.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ObjectId": {
        "type": "object",
        "properties": {}
      },
      "CreateCustomFieldDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the custom field"
          },
          "type": {
            "type": "string",
            "description": "The type of the custom field"
          },
          "options": {
            "description": "The options for the custom field",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "position": {
            "type": "number",
            "description": "The position of the custom field"
          },
          "schema": {
            "type": "string",
            "description": "The schema of the custom field"
          },
          "ownerId": {
            "type": "string",
            "description": "The owner ID of the custom field"
          },
          "schemaIds": {
            "description": "The schema IDs associated with the custom field",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        },
        "required": [
          "name",
          "type",
          "schema",
          "ownerId"
        ]
      },
      "UpdateCustomFieldDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the custom field"
          },
          "type": {
            "type": "string",
            "description": "The type of the custom field"
          },
          "options": {
            "description": "The options for the custom field",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "position": {
            "type": "number",
            "description": "The position of the custom field"
          },
          "schema": {
            "type": "string",
            "description": "The schema of the custom field"
          },
          "ownerId": {
            "type": "string",
            "description": "The owner ID of the custom field"
          },
          "schemaIds": {
            "description": "The schema IDs associated with the custom field",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        }
      },
...
    }
  }
}